checkIfClosed
